requirements.txt
fastapi
uvicorn[standard]
python-multipart
pdfminer.six
python-docx
fpdf2
spacy
sentence-transformers
transformers
torch
faiss-cpu
pydantic
requests
pytest
python-dotenv
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
PyJWT
bcrypt
